+++ Modulo13/Liger/main.cpp
@@ -7,9 +7,8 @@ int main() {
 
     Liger lg(100);
 
-    int weight = lg.getWeight();
+    float weight = lg.getWeight();
 
-    
     cout << weight << endl;
     
     return 0;
